// web/src/api/generated/models/SpaceConfigurationView.ts
export interface SpaceConfigurationView {
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    canManageSpaceConfiguration: boolean;
    /**
     * 
     * @type {string}
     * @memberof SpaceConfigurationView
     */
    dailyContextTimezone: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    dailyQuestionsEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    energyCheckInEnabled: boolean;
    /**
     * 
     * @type {DailyCheckInVisibilityMode}
     * @memberof SpaceConfigurationView
     */
    energyVisibilityMode: DailyCheckInVisibilityMode;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    loveNotesEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    sharedAchievementsEnabled: boolean;
    /**
     * 
     * @type {string}
     * @memberof SpaceConfigurationView
     */
    spaceId: string;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    supportGesturesEnabled: boolean;
    /**
     * 
     * @type {number}
     * @memberof SpaceConfigurationView
     */
    version: number;
    /**
     * 
     * @type {boolean}
     * @memberof SpaceConfigurationView
     */
    vibeCheckEnabled: boolean;
    /**
     * 
     * @type {DailyCheckInVisibilityMode}
     * @memberof SpaceConfigurationView
     */
    vibeVisibilityMode: DailyCheckInVisibilityMode;
}